+++ Philosophers/get_time.c
@@ -6,11 +6,12 @@
 /*   By: ael-asri <ael-asri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/24 20:37:59 by ael-asri          #+#    #+#             */
-/*   Updated: 2022/04/02 21:19:06 by ael-asri         ###   ########.fr       */
+/*   Updated: 2022/04/16 01:06:30 by ael-asri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philosophers.h"
+//#include "philosophers.h"
+#include "_bonus/philosophers_bonus.h"
 /*
 int	get_time()
 {
@@ -37,7 +38,7 @@ long	ft_time(void)
 	return (res);
 }
 
-long	get_time()
+unsigned long long	get_time()
 {
 	struct timeval	current_time;
 //	int				time;
